Move RunConfig class to own file
